+++ src/lib/components/properties-menu/models/property-list-item.model.ts
@@ -2,11 +2,33 @@ import { IconDefinition } from '@fortawesome/angular-fontawesome';
 
 import { PropertyBadge, PropertyBadgeParameters, PropertySummaryRow } from './property-summary-row.model';
 
+export interface PropertyListItemActionParameters {
+    icon: IconDefinition;
+    key: string;
+
+    label?: string;
+}
+
+export class PropertyListItemAction {
+    icon: IconDefinition;
+    key: string;
+
+    label?: string;
+
+    constructor({ icon, key, label }: PropertyListItemActionParameters) {
+        this.icon = icon;
+        this.key = key;
+        this.label = label;
+    }
+}
+
 export interface PropertyListItemParameters {
     id: string;
 
+    actions?: PropertyListItemActionParameters[];
     badges?: PropertyBadgeParameters[];
     body?: PropertySummaryRow[];
+    copyValue?: string;
     description?: string;
     disabled?: boolean;
     expanded?: boolean;
@@ -20,6 +42,7 @@ export interface PropertyListItemParameters {
 }
 
 export class PropertyListItem {
+    actions: PropertyListItemAction[];
     badges: PropertyBadge[];
     disabled: boolean;
     expanded: boolean;
@@ -30,14 +53,17 @@ export class PropertyListItem {
     removable: boolean;
 
     body?: PropertySummaryRow[];
+    copyValue?: string;
     description?: string;
     icon?: IconDefinition;
     iconClasses?: string;
     labelParameters?: Record<string, unknown>;
 
     constructor({
+        actions = [],
         badges = [],
         body,
+        copyValue,
         description,
         disabled = false,
         expanded = false,
@@ -50,8 +76,12 @@ export class PropertyListItem {
         metadata = {},
         removable = false
     }: PropertyListItemParameters) {
+        this.actions = actions.map(action =>
+            action instanceof PropertyListItemAction ? action : new PropertyListItemAction(action)
+        );
         this.badges = badges.map(badge => (badge instanceof PropertyBadge ? badge : new PropertyBadge(badge)));
         this.body = body;
+        this.copyValue = copyValue;
         this.description = description;
         this.disabled = disabled;
         this.expanded = expanded;